chore: refactor debugger code
